Move the tests into the "tests" directory. Add additional tests for the `state()` method
