+++ src/lib.rs
@@ -4,4 +4,4 @@ mod subscription;
 
 pub use reducer::Reducer;
 pub use store::Store;
-pub use subscription::Subscription;
+pub use subscription::{Subscription, UnsubscribeError};
